Rename reactiveContext._id to reactiveContext.id
The meteor templating system creates _id as the 'id' of an object, however reactiveContext._id is not an id and is not unique, rather it represents the _id of the object being edited, known as the 'item'. I've renamed this to id which makes it clear to the templating system that it cannot rely on that property to distinguish between reactiveContext instances.